+++ base/deprecated.jl
@@ -1705,6 +1705,9 @@ export hex2num
 # issue #17886
 # deprecations for filter[!] with 2-arg functions are in associative.jl
 
+# PR #23066
+@deprecate cfunction(f, r, a::Tuple) cfunction(f, r, Tuple{a...})
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations